+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:av_media_player/index.dart';
+import 'package:av_media_player/widget.dart';
 
 void main() => runApp(const MyApp());
 
@@ -11,30 +11,29 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  AVMediaPlayer? _player;
+  var _loading = true;
 
   @override
   Widget build(BuildContext context) => MaterialApp(
         home: Scaffold(
           appBar: AppBar(
-            title: const Text('AV Media Player example app'),
+            title: const Text('av_media_player simple example'),
           ),
-          body: Stack(
-            alignment: Alignment.center,
-            children: [
-              AVMediaView(
-                initSource:
-                    'https://devstreaming-cdn.apple.com/videos/streaming/examples/bipbop_4x3/bipbop_4x3_variant.m3u8',
-                initLooping: true,
-                initAutoPlay: true,
-                onCreated: (player) {
-                  _player = player;
-                  player.loading.addListener(() => setState(() {}));
-                },
-              ),
-              if (_player?.loading.value ?? true)
-                const CircularProgressIndicator(),
-            ],
+          body: SizedBox.expand(
+            child: Stack(
+              alignment: Alignment.center,
+              children: [
+                AVMediaView(
+                  initSource:
+                      'https://devstreaming-cdn.apple.com/videos/streaming/examples/bipbop_4x3/bipbop_4x3_variant.m3u8',
+                  initLooping: true,
+                  initAutoPlay: true,
+                  onCreated: (player) => player.loading.addListener(
+                      () => setState(() => _loading = player.loading.value)),
+                ),
+                if (_loading) const CircularProgressIndicator(),
+              ],
+            ),
           ),
         ),
       );